implemented new nodes needed for complete display

+++ src/parserModel/nodes/display/SetBackgroundNode.java
@@ -8,7 +8,7 @@ public class SetBackgroundNode extends UnaryOperationNode {
     @Override
     public double execute(TurtleContext context) {
         double backgroundIndex = myArgumentNode.execute(context);
-        context.addToQueue(new SetBackgroundExecutable(backgroundIndex));
+        context.addToQueue(new SetBackgroundExecutable((int)backgroundIndex));
         return backgroundIndex;
     }
 }
